Add a startup hook
