+++ pkg/lookuptable/lookuptable.go
@@ -1,0 +1,7 @@
+package lookuptable
+
+type LookupTable struct {
+	StartupHook func()
+}
+
+var GlobalLookupTable LookupTable
